+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- A command for CLI to read source text from file
+
 ### Changed
 - Fix display of languages in the terminal
 - Show message if not specified --speak-source or --speak-translation options in --audio-only mode